+++ src/main/java/com/kinoxp/service/UserService.java
@@ -28,7 +28,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public void deleteUser(int id) {
+    public void deleteUser(long id) {
         userRepository.deleteById(id);
     }
 }
